test(site-explorer): split focused integration tests into crate
Move selected site explorer tests out of api-core into dedicated
site-explorer integration test files, and share common test environment
setup for site-explorer/power-shelf tests. Add the dev uuid dependency
needed by migrated metrics tests.
