+++ crates/api-core/src/tests/site_explorer.rs
@@ -329,79 +329,6 @@ async fn test_site_explorer_default_pause_ingestion_and_poweron(
 }
 
 #[sqlx_test]
-async fn test_handle_redfish_error_powers_on_machine(
-    pool: PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-
-    let mut machine = FakeMachine::new("6a:6b:6c:6d:6e:70", "Vendor1");
-    machine.discover_dhcp(&env).await?;
-    let bmc_ip: IpAddr = machine.ip.parse()?;
-
-    let mut txn = env.pool.begin().await?;
-    db::expected_machine::create(
-        &mut txn,
-        ExpectedMachine {
-            id: None,
-            bmc_mac_address: machine.mac,
-            data: ExpectedMachineData {
-                serial_number: "host-needs-power-on".to_string(),
-                ..Default::default()
-            },
-        },
-    )
-    .await?;
-    txn.commit().await?;
-
-    let endpoint_explorer = Arc::new(MockEndpointExplorer::default());
-    endpoint_explorer.insert_endpoint_result(
-        bmc_ip,
-        Err(EndpointExplorationError::RedfishError {
-            details: "transient redfish failure".to_string(),
-            response_body: None,
-            response_code: Some(500),
-        }),
-    );
-    endpoint_explorer
-        .power_states
-        .lock()
-        .unwrap()
-        .insert(bmc_ip, libredfish::PowerState::Off);
-
-    let explorer_config = SiteExplorerConfig {
-        enabled: Arc::new(true.into()),
-        explorations_per_run: 1,
-        concurrent_explorations: 1,
-        run_interval: std::time::Duration::from_secs(1),
-        create_machines: Arc::new(true.into()),
-        ..Default::default()
-    };
-    let explorer = env.new_site_explorer(explorer_config, &endpoint_explorer);
-
-    explorer.run_single_iteration().await?;
-
-    {
-        let calls = endpoint_explorer
-            .redfish_power_control_calls
-            .lock()
-            .unwrap();
-        assert_eq!(
-            calls.as_slice(),
-            &[(
-                std::net::SocketAddr::new(bmc_ip, 443),
-                libredfish::SystemPowerControl::On
-            )]
-        );
-    }
-
-    let mut txn = env.pool.begin().await?;
-    let endpoints = db::explored_endpoints::find_all_by_ip(bmc_ip, &mut txn).await?;
-    txn.commit().await?;
-    assert_eq!(endpoints.len(), 1, "expected one explored endpoint");
-    Ok(())
-}
-
-#[sqlx_test]
 async fn test_site_explorer_main(pool: PgPool) -> Result<(), Box<dyn std::error::Error>> {
     let env = common::api_fixtures::create_test_env(pool.clone()).await;
     let underlay_segment = env.underlay_segment.unwrap();
@@ -787,250 +714,6 @@ async fn test_site_explorer_main(pool: PgPool) -> Result<(), Box<dyn std::error:
     );
 
     txn.commit().await?;
-    Ok(())
-}
-
-/// Strict ingestion gate: a host whose BMC reports no DPU PCIe devices
-/// and whose `ExpectedMachine` does not declare `NoDpu` is skipped (with
-/// a warning + a `NoDpuReportedByHost` pairing-blocker metric) rather
-/// than ingested. Operators must explicitly opt in to zero-DPU.
-#[sqlx_test]
-async fn test_site_explorer_skips_unexpected_zero_dpu_host(
-    pool: PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-
-    let mut machine = FakeMachine::new("AA:AB:AC:AD:AA:11", "Vendor1");
-    machine.discover_dhcp(&env).await?;
-
-    // expected_machine WITHOUT a NoDpu declaration -- the host is
-    // "expected to have DPUs" by default.
-    let mut txn = env.pool.begin().await?;
-    db::expected_machine::create(
-        &mut txn,
-        ExpectedMachine {
-            id: None,
-            bmc_mac_address: machine.mac,
-            data: ExpectedMachineData {
-                serial_number: "host-expected-dpus-but-has-none".to_string(),
-                ..Default::default()
-            },
-        },
-    )
-    .await?;
-    txn.commit().await?;
-
-    // BMC report with no PCIe devices / no chassis -- the gate sees
-    // zero DPUs.
-    let endpoint_explorer = Arc::new(MockEndpointExplorer::default());
-    endpoint_explorer.insert_endpoint_results(vec![(
-        machine.ip.parse().unwrap(),
-        Ok(EndpointExplorationReport {
-            endpoint_type: EndpointType::Bmc,
-            vendor: Some(bmc_vendor::BMCVendor::Lenovo),
-            systems: vec![ComputerSystem {
-                serial_number: Some("0123456789".to_string()),
-                ..Default::default()
-            }],
-            ..Default::default()
-        }),
-    )]);
-
-    let explorer_config = SiteExplorerConfig {
-        enabled: Arc::new(true.into()),
-        explorations_per_run: 1,
-        concurrent_explorations: 1,
-        run_interval: std::time::Duration::from_secs(1),
-        create_machines: Arc::new(true.into()),
-        ..Default::default()
-    };
-    let explorer = env.new_site_explorer(explorer_config, &endpoint_explorer);
-    let test_meter = &env.test_meter;
-
-    // First iteration populates `explored_endpoints`; second runs
-    // `identify_managed_hosts` after preingestion is complete.
-    explorer.run_single_iteration().await.unwrap();
-    let mut txn = env.pool.begin().await?;
-    db::explored_endpoints::set_preingestion_complete(machine.ip.parse().unwrap(), &mut txn)
-        .await?;
-    txn.commit().await?;
-    explorer.run_single_iteration().await.unwrap();
-
-    // No managed host should have been identified.
-    let explored_managed_hosts = db::explored_managed_host::find_all(&env.pool).await?;
-    assert!(
-        explored_managed_hosts.is_empty(),
-        "strict gate should refuse to ingest a zero-DPU host without a `NoDpu` declaration, got {:?}",
-        explored_managed_hosts,
-    );
-
-    assert_eq!(
-        test_meter
-            .formatted_metric("carbide_site_exploration_identified_managed_hosts_count")
-            .unwrap(),
-        "0"
-    );
-
-    // The pairing-blocker metric should have ticked for `NoDpuReportedByHost`.
-    let blocker_metric = test_meter
-        .formatted_metric("carbide_host_dpu_pairing_blockers_count")
-        .expect("expected `carbide_host_dpu_pairing_blockers_count` to be emitted");
-    assert!(
-        blocker_metric.contains("no_dpu_reported_by_host"),
-        "expected pairing-blocker metric to mention `no_dpu_reported_by_host`, got {blocker_metric}",
-    );
-
-    Ok(())
-}
-
-/// Companion to `test_site_explorer_skips_unexpected_zero_dpu_host`: when
-/// the operator explicitly declares `dpu_mode = "nic_mode"`, a host whose
-/// BMC reports zero usable DPU PCIe devices (because anything that is a
-/// BlueField has been stripped as "DPU in NIC mode") should be ingested as
-/// a zero-DPU managed host -- the operator has already opted into "treat
-/// as zero-DPU" semantics by declaring NicMode.
-#[sqlx_test]
-async fn test_site_explorer_ingests_nic_mode_host_with_no_observed_dpus(
-    pool: PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-
-    let mut machine = FakeMachine::new("AA:AB:AC:AD:AA:22", "Vendor1");
-    machine.discover_dhcp(&env).await?;
-
-    let mut txn = env.pool.begin().await?;
-    db::expected_machine::create(
-        &mut txn,
-        ExpectedMachine {
-            id: None,
-            bmc_mac_address: machine.mac,
-            data: ExpectedMachineData {
-                serial_number: "host-nic-mode-no-observed-dpus".to_string(),
-                dpu_mode: model::expected_machine::DpuMode::NicMode,
-                ..Default::default()
-            },
-        },
-    )
-    .await?;
-    txn.commit().await?;
-
-    let endpoint_explorer = Arc::new(MockEndpointExplorer::default());
-    endpoint_explorer.insert_endpoint_results(vec![(
-        machine.ip.parse().unwrap(),
-        Ok(EndpointExplorationReport {
-            endpoint_type: EndpointType::Bmc,
-            vendor: Some(bmc_vendor::BMCVendor::Lenovo),
-            systems: vec![ComputerSystem {
-                serial_number: Some("0123456789".to_string()),
-                ..Default::default()
-            }],
-            ..Default::default()
-        }),
-    )]);
-
-    let explorer_config = SiteExplorerConfig {
-        enabled: Arc::new(true.into()),
-        explorations_per_run: 1,
-        concurrent_explorations: 1,
-        run_interval: std::time::Duration::from_secs(1),
-        create_machines: Arc::new(true.into()),
-        ..Default::default()
-    };
-    let explorer = env.new_site_explorer(explorer_config, &endpoint_explorer);
-
-    explorer.run_single_iteration().await.unwrap();
-    let mut txn = env.pool.begin().await?;
-    db::explored_endpoints::set_preingestion_complete(machine.ip.parse().unwrap(), &mut txn)
-        .await?;
-    txn.commit().await?;
-    explorer.run_single_iteration().await.unwrap();
-
-    let explored_managed_hosts = db::explored_managed_host::find_all(&env.pool).await?;
-    assert_eq!(
-        explored_managed_hosts.len(),
-        1,
-        "NicMode declaration should let the host through the strict gate even with zero observed DPUs",
-    );
-    assert!(
-        explored_managed_hosts[0].dpus.is_empty(),
-        "NicMode hosts ingest with an empty `dpus` vector",
-    );
-
-    Ok(())
-}
-
-/// Third member of the zero-DPU triad (alongside the `DpuMode::DpuMode`
-/// skip test and the `DpuMode::NicMode` ingest test): a host explicitly
-/// declared `dpu_mode = "no_dpu"` ingests as a zero-DPU managed host. The
-/// `NoDpu` fast-path in `identify_managed_hosts` short-circuits before any
-/// DPU PCIe enumeration, so this holds regardless of what the BMC reports.
-#[sqlx_test]
-async fn test_site_explorer_ingests_no_dpu_host(
-    pool: PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-
-    let mut machine = FakeMachine::new("AA:AB:AC:AD:AA:33", "Vendor1");
-    machine.discover_dhcp(&env).await?;
-
-    let mut txn = env.pool.begin().await?;
-    db::expected_machine::create(
-        &mut txn,
-        ExpectedMachine {
-            id: None,
-            bmc_mac_address: machine.mac,
-            data: ExpectedMachineData {
-                serial_number: "host-no-dpu-declared".to_string(),
-                dpu_mode: model::expected_machine::DpuMode::NoDpu,
-                ..Default::default()
-            },
-        },
-    )
-    .await?;
-    txn.commit().await?;
-
-    let endpoint_explorer = Arc::new(MockEndpointExplorer::default());
-    endpoint_explorer.insert_endpoint_results(vec![(
-        machine.ip.parse().unwrap(),
-        Ok(EndpointExplorationReport {
-            endpoint_type: EndpointType::Bmc,
-            vendor: Some(bmc_vendor::BMCVendor::Lenovo),
-            systems: vec![ComputerSystem {
-                serial_number: Some("0123456789".to_string()),
-                ..Default::default()
-            }],
-            ..Default::default()
-        }),
-    )]);
-
-    let explorer_config = SiteExplorerConfig {
-        enabled: Arc::new(true.into()),
-        explorations_per_run: 1,
-        concurrent_explorations: 1,
-        run_interval: std::time::Duration::from_secs(1),
-        create_machines: Arc::new(true.into()),
-        ..Default::default()
-    };
-    let explorer = env.new_site_explorer(explorer_config, &endpoint_explorer);
-
-    explorer.run_single_iteration().await.unwrap();
-    let mut txn = env.pool.begin().await?;
-    db::explored_endpoints::set_preingestion_complete(machine.ip.parse().unwrap(), &mut txn)
-        .await?;
-    txn.commit().await?;
-    explorer.run_single_iteration().await.unwrap();
-
-    let explored_managed_hosts = db::explored_managed_host::find_all(&env.pool).await?;
-    assert_eq!(
-        explored_managed_hosts.len(),
-        1,
-        "NoDpu declaration should ingest the host as zero-DPU",
-    );
-    assert!(
-        explored_managed_hosts[0].dpus.is_empty(),
-        "NoDpu hosts ingest with an empty `dpus` vector",
-    );
-
     Ok(())
 }
 
@@ -2471,76 +2154,6 @@ async fn test_site_explorer_fixtures_zerodpu_dhcp_before_site_explorer(
 }
 
 #[sqlx_test]
-async fn test_site_explorer_unknown_vendor(pool: PgPool) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-    let underlay_segment = env.underlay_segment.unwrap();
-
-    let mut machine = FakeMachine::new("B8:3F:D2:90:97:A7", "Vendor1");
-    machine.discover_dhcp(&env).await?;
-
-    let mut txn = env.pool.begin().await?;
-    assert_eq!(
-        db::machine_interface::count_by_segment_id(&mut txn, &underlay_segment)
-            .await
-            .unwrap(),
-        1
-    );
-    txn.commit().await.unwrap();
-
-    let endpoint_explorer = Arc::new(MockEndpointExplorer::default());
-
-    endpoint_explorer.insert_endpoint_result(
-        machine.ip.parse().unwrap(),
-        Err(EndpointExplorationError::UnsupportedVendor {
-            vendor: "Unknown".to_string(),
-        }),
-    );
-
-    let explorer_config = SiteExplorerConfig {
-        enabled: Arc::new(true.into()),
-        explorations_per_run: 2,
-        concurrent_explorations: 1,
-        run_interval: std::time::Duration::from_secs(1),
-        create_machines: Arc::new(true.into()),
-        allocate_secondary_vtep_ip: true,
-        create_power_shelves: Arc::new(true.into()),
-        explore_power_shelves_from_static_ip: Arc::new(true.into()),
-        power_shelves_created_per_run: 1,
-        create_switches: Arc::new(true.into()),
-        switches_created_per_run: 1,
-        ..Default::default()
-    };
-    let explorer = env.new_site_explorer(explorer_config, &endpoint_explorer);
-
-    explorer.run_single_iteration().await.unwrap();
-    // Since we configured a limit of 2 entries, we should have those 2 results now
-    let mut txn = env.pool.begin().await?;
-    let explored = db::explored_endpoints::find_all(txn.as_mut())
-        .await
-        .unwrap();
-    txn.commit().await?;
-    assert_eq!(explored.len(), 1);
-    let report = &explored[0];
-    assert_eq!(report.report_version.version_nr(), 1);
-    assert_eq!(
-        report.report.last_exploration_error,
-        Some(EndpointExplorationError::UnsupportedVendor {
-            vendor: "Unknown".to_string(),
-        })
-    );
-
-    let guard = endpoint_explorer.reports.lock().unwrap();
-    let res = guard.get(&report.address).unwrap().as_ref();
-    assert!(res.is_err());
-    assert_eq!(
-        res.unwrap_err(),
-        report.report.last_exploration_error.as_ref().unwrap()
-    );
-
-    Ok(())
-}
-
-#[sqlx_test]
 async fn test_delete_explored_endpoint(pool: PgPool) -> Result<(), Box<dyn std::error::Error>> {
     let env = common::api_fixtures::create_test_env(pool.clone()).await;
 
@@ -2811,298 +2424,6 @@ async fn test_machine_creation_with_sku(pool: PgPool) -> Result<(), Box<dyn std:
     assert!(!expected_metrics.is_empty());
     // The SKU "Sku1" has device_type=None, so it should be counted with device_type="unknown"
     assert!(expected_metrics.contains_key("{device_type=\"unknown\",sku_id=\"Sku1\"}"));
-
-    Ok(())
-}
-
-#[sqlx_test]
-async fn test_expected_machine_device_type_metrics(
-    pool: PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-
-    let test_sku_gpu_id = format!("test-sku-gpu-{}", uuid::Uuid::new_v4());
-    let test_sku_no_type_id = format!("test-sku-no-type-{}", uuid::Uuid::new_v4());
-    const EXPECTED_MACHINE_1_MAC: &str = "AA:BB:CC:DD:EE:01";
-    const EXPECTED_MACHINE_2_MAC: &str = "AA:BB:CC:DD:EE:02";
-    const EXPECTED_MACHINE_3_MAC: &str = "AA:BB:CC:DD:EE:03";
-
-    // Create fake machines with network interfaces so they can be discovered
-    let mut machines = vec![
-        FakeMachine::new(EXPECTED_MACHINE_1_MAC, "Vendor1"),
-        FakeMachine::new(EXPECTED_MACHINE_2_MAC, "Vendor2"),
-        FakeMachine::new(EXPECTED_MACHINE_3_MAC, "Vendor3"),
-    ];
-    machines.discover_dhcp(&env).await?;
-
-    // Create test SKUs in database
-    let mut txn = env.pool.begin().await?;
-
-    let test_sku_with_device_type = model::sku::Sku {
-        schema_version: CURRENT_SKU_VERSION,
-        id: test_sku_gpu_id.clone(),
-        description: "Test GPU SKU".to_string(),
-        created: chrono::Utc::now(),
-        components: model::sku::SkuComponents {
-            chassis: model::sku::SkuComponentChassis {
-                vendor: format!("test_vendor_gpu_{}", uuid::Uuid::new_v4()),
-                model: format!("test_model_gpu_{}", uuid::Uuid::new_v4()),
-                architecture: "x86_64".to_string(),
-            },
-            cpus: vec![],
-            gpus: vec![],
-            memory: vec![],
-            infiniband_devices: vec![],
-            storage: vec![],
-            tpm: None,
-        },
-        device_type: Some("gpu".to_string()),
-    };
-
-    let test_sku_without_device_type = model::sku::Sku {
-        schema_version: CURRENT_SKU_VERSION,
-        id: test_sku_no_type_id.clone(),
-        description: "Test SKU without device type".to_string(),
-        created: chrono::Utc::now(),
-        components: model::sku::SkuComponents {
-            chassis: model::sku::SkuComponentChassis {
-                vendor: format!("test_vendor_no_type_{}", uuid::Uuid::new_v4()),
-                model: format!("test_model_no_type_{}", uuid::Uuid::new_v4()),
-                architecture: "x86_64".to_string(),
-            },
-            cpus: vec![],
-            gpus: vec![],
-            memory: vec![],
-            infiniband_devices: vec![],
-            storage: vec![],
-            tpm: None,
-        },
-        device_type: None,
-    };
-
-    db::sku::create(&mut txn, &test_sku_with_device_type).await?;
-    db::sku::create(&mut txn, &test_sku_without_device_type).await?;
-
-    // Create expected machines with different SKU configurations
-    db::expected_machine::create(
-        &mut txn,
-        ExpectedMachine {
-            id: None,
-            bmc_mac_address: EXPECTED_MACHINE_1_MAC.parse().unwrap(),
-            data: ExpectedMachineData {
-                bmc_username: "user1".to_string(),
-                bmc_password: "pass1".to_string(),
-                serial_number: "serial1".to_string(),
-                fallback_dpu_serial_numbers: vec![],
-                metadata: Metadata::new_with_default_name(),
-                sku_id: Some(test_sku_gpu_id.clone()),
-                default_pause_ingestion_and_poweron: None,
-                host_nics: vec![],
-                rack_id: None,
-                dpf_enabled: Some(true),
-                bmc_ip_address: None,
-                bmc_retain_credentials: None,
-                dpu_mode: Default::default(),
-                host_lifecycle_profile: Default::default(),
-            },
-        },
-    )
-    .await?;
-
-    db::expected_machine::create(
-        &mut txn,
-        ExpectedMachine {
-            id: None,
-            bmc_mac_address: EXPECTED_MACHINE_2_MAC.parse().unwrap(),
-            data: ExpectedMachineData {
-                bmc_username: "user2".to_string(),
-                bmc_password: "pass2".to_string(),
-                serial_number: "serial2".to_string(),
-                fallback_dpu_serial_numbers: vec![],
-                metadata: Metadata::new_with_default_name(),
-                sku_id: Some(test_sku_no_type_id.clone()),
-                default_pause_ingestion_and_poweron: None,
-                host_nics: vec![],
-                rack_id: None,
-                dpf_enabled: Some(true),
-                bmc_ip_address: None,
-                bmc_retain_credentials: None,
-                dpu_mode: Default::default(),
-                host_lifecycle_profile: Default::default(),
-            },
-        },
-    )
-    .await?;
-
-    db::expected_machine::create(
-        &mut txn,
-        ExpectedMachine {
-            id: None,
-            bmc_mac_address: EXPECTED_MACHINE_3_MAC.parse().unwrap(),
-            data: ExpectedMachineData {
-                bmc_username: "user3".to_string(),
-                bmc_password: "pass3".to_string(),
-                serial_number: "serial3".to_string(),
-                fallback_dpu_serial_numbers: vec![],
-                metadata: Metadata::new_with_default_name(),
-                sku_id: None, // No SKU
-                default_pause_ingestion_and_poweron: None,
-                host_nics: vec![],
-                rack_id: None,
-                dpf_enabled: Some(true),
-                bmc_ip_address: None,
-                bmc_retain_credentials: None,
-                dpu_mode: Default::default(),
-                host_lifecycle_profile: Default::default(),
-            },
-        },
-    )
-    .await?;
-
-    txn.commit().await?;
-
-    // Set up endpoint explorer with mock results for our machines
-    let endpoint_explorer = Arc::new(MockEndpointExplorer::default());
-
-    // Mock exploration results for each machine
-    endpoint_explorer.insert_endpoint_results(vec![
-        (
-            machines[0].ip.parse().unwrap(),
-            Ok(EndpointExplorationReport {
-                endpoint_type: EndpointType::Bmc,
-                last_exploration_error: None,
-                last_exploration_latency: Some(std::time::Duration::from_millis(100)),
-                vendor: Some(bmc_vendor::BMCVendor::Dell),
-                managers: vec![],
-                systems: vec![],
-                chassis: vec![],
-                service: vec![],
-                machine_id: None,
-                versions: std::collections::HashMap::new(),
-                model: Some("test-model".to_string()),
-                machine_setup_status: None,
-                secure_boot_status: None,
-                lockdown_status: None,
-                power_shelf_id: None,
-                switch_id: None,
-                compute_tray_index: None,
-                physical_slot_number: None,
-                revision_id: None,
-                topology_id: None,
-                remediation_error: None,
-            }),
-        ),
-        (
-            machines[1].ip.parse().unwrap(),
-            Ok(EndpointExplorationReport {
-                endpoint_type: EndpointType::Bmc,
-                last_exploration_error: None,
-                last_exploration_latency: Some(std::time::Duration::from_millis(100)),
-                vendor: Some(bmc_vendor::BMCVendor::Nvidia),
-                managers: vec![],
-                systems: vec![],
-                chassis: vec![],
-                service: vec![],
-                machine_id: None,
-                versions: std::collections::HashMap::new(),
-                model: Some("test-model".to_string()),
-                machine_setup_status: None,
-                secure_boot_status: None,
-                lockdown_status: None,
-                power_shelf_id: None,
-                switch_id: None,
-                compute_tray_index: None,
-                physical_slot_number: None,
-                revision_id: None,
-                topology_id: None,
-                remediation_error: None,
-            }),
-        ),
-        (
-            machines[2].ip.parse().unwrap(),
-            Ok(EndpointExplorationReport {
-                endpoint_type: EndpointType::Bmc,
-                last_exploration_error: None,
-                last_exploration_latency: Some(std::time::Duration::from_millis(100)),
-                vendor: Some(bmc_vendor::BMCVendor::Supermicro),
-                managers: vec![],
-                systems: vec![],
-                chassis: vec![],
-                service: vec![],
-                machine_id: None,
-                versions: std::collections::HashMap::new(),
-                model: Some("test-model".to_string()),
-                machine_setup_status: None,
-                secure_boot_status: None,
-                lockdown_status: None,
-                power_shelf_id: None,
-                switch_id: None,
-                compute_tray_index: None,
-                physical_slot_number: None,
-                revision_id: None,
-                topology_id: None,
-                remediation_error: None,
-            }),
-        ),
-    ]);
-
-    let explorer_config = SiteExplorerConfig {
-        enabled: Arc::new(true.into()),
-        explorations_per_run: 3, // Explore our 3 machines
-        concurrent_explorations: 1,
-        run_interval: std::time::Duration::from_secs(1),
-        create_machines: Arc::new(false.into()),
-        allocate_secondary_vtep_ip: true,
-        create_power_shelves: Arc::new(true.into()),
-        explore_power_shelves_from_static_ip: Arc::new(true.into()),
-        power_shelves_created_per_run: 1,
-        create_switches: Arc::new(true.into()),
-        switches_created_per_run: 1,
-        ..Default::default()
-    };
-
-    let explorer = env.new_site_explorer(explorer_config, &endpoint_explorer);
-    let test_meter = &env.test_meter;
-
-    // Run site explorer to collect metrics
-    explorer.run_single_iteration().await.unwrap();
-
-    // Verify expected machines SKU count metrics
-    let device_type_metrics: HashMap<String, String> = test_meter
-        .parsed_metrics("carbide_site_exploration_expected_machines_sku_count")
-        .into_iter()
-        .collect();
-
-    assert!(!device_type_metrics.is_empty());
-
-    // Expected machines metrics are now recorded based on both SKU ID and device type
-    // Now that we properly set device_type using update_metadata:
-    // - 1 machine with GPU SKU -> sku_id=test_sku_gpu_id, device_type="gpu"
-    // - 1 machine with no device_type SKU -> sku_id=test_sku_no_type_id, device_type="unknown"
-    // - 1 machine with no SKU -> sku_id="unknown", device_type="unknown"
-
-    // Check machine with GPU SKU
-    let gpu_sku_key = format!("{{device_type=\"gpu\",sku_id=\"{test_sku_gpu_id}\"}}");
-    assert_eq!(device_type_metrics.get(&gpu_sku_key).unwrap(), "1");
-
-    // Check machine with SKU but no device type
-    let no_type_sku_key = format!("{{device_type=\"unknown\",sku_id=\"{test_sku_no_type_id}\"}}");
-    assert_eq!(device_type_metrics.get(&no_type_sku_key).unwrap(), "1");
-
-    // Check machine with no SKU
-    assert_eq!(
-        device_type_metrics
-            .get("{device_type=\"unknown\",sku_id=\"unknown\"}")
-            .unwrap(),
-        "1"
-    );
-
-    // Verify total count by summing all device types
-    let total_count: u32 = device_type_metrics
-        .values()
-        .map(|v| v.parse::<u32>().unwrap())
-        .sum();
-    assert_eq!(total_count, 3);
 
     Ok(())
 }
@@ -3913,396 +3234,6 @@ async fn switch_treats_na_chassis_serial_as_missing(
     assert!(
         ids.is_empty(),
         "no switch row must be inserted when chassis serial is NA"
-    );
-
-    Ok(())
-}
-
-/// Site-explorer reconciles every `expected_*` row's configured static IPs into
-/// `machine_interface` rows by calling `try_preallocate_one` per static IP during
-/// `update_explored_endpoints`. This test drives the same per-row materialization directly,
-/// covering the static-assignments-segment counterpart to the DHCP `discover()` recovery hook:
-/// devices whose IP lives outside any Carbide-managed network never reach `discover()`, so this
-/// per-row preallocation is what gets their rows onto the books.
-#[sqlx_test]
-async fn test_site_explorer_reconcile_creates_missing_preallocations(
-    pool: PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-
-    let machine_bmc_mac: MacAddress = "AA:BB:CC:DD:E0:01".parse().unwrap();
-    let machine_bmc_ip: IpAddr = "10.99.0.10".parse().unwrap();
-    let switch_bmc_mac: MacAddress = "AA:BB:CC:DD:E0:02".parse().unwrap();
-    let switch_bmc_ip: IpAddr = "10.99.0.11".parse().unwrap();
-    let power_shelf_bmc_mac: MacAddress = "AA:BB:CC:DD:E0:03".parse().unwrap();
-    let power_shelf_bmc_ip: IpAddr = "10.99.0.12".parse().unwrap();
-
-    // Seed each expected_* row WITHOUT a corresponding machine_interface. The gRPC `add`
-    // handlers don't preallocate inline; site-explorer's reconciliation pass is what
-    // materializes the rows.
-    let mut txn = env.pool.begin().await?;
-    db::expected_machine::create(
-        &mut txn,
-        ExpectedMachine {
-            id: None,
-            bmc_mac_address: machine_bmc_mac,
-            data: ExpectedMachineData {
-                serial_number: "reconcile-m-001".to_string(),
-                bmc_ip_address: Some(machine_bmc_ip),
-                ..Default::default()
-            },
-        },
-    )
-    .await?;
-    db::expected_switch::create(
-        &mut txn,
-        model::expected_switch::ExpectedSwitch {
-            expected_switch_id: None,
-            bmc_mac_address: switch_bmc_mac,
-            nvos_mac_addresses: vec![],
-            bmc_username: "ADMIN".into(),
-            serial_number: "reconcile-sw-001".into(),
-            bmc_password: "PASS".into(),
-            nvos_username: None,
-            nvos_password: None,
-            bmc_ip_address: Some(switch_bmc_ip),
-            nvos_ip_address: None,
-            metadata: Metadata::default(),
-            rack_id: None,
-            bmc_retain_credentials: None,
-        },
-    )
-    .await?;
-    db::expected_power_shelf::create(
-        &mut txn,
-        model::expected_power_shelf::ExpectedPowerShelf {
-            expected_power_shelf_id: None,
-            bmc_mac_address: power_shelf_bmc_mac,
-            bmc_username: "ADMIN".into(),
-            serial_number: "reconcile-ps-001".into(),
-            bmc_password: "PASS".into(),
-            bmc_ip_address: Some(power_shelf_bmc_ip),
-            metadata: Metadata::default(),
-            rack_id: None,
-            bmc_retain_credentials: None,
-        },
-    )
-    .await?;
-    txn.commit().await?;
-
-    // Baseline: no interfaces yet.
-    let mut txn = env.pool.begin().await?;
-    for mac in [machine_bmc_mac, switch_bmc_mac, power_shelf_bmc_mac] {
-        let before = db::machine_interface::find_by_mac_address(&mut *txn, mac).await?;
-        assert!(
-            before.is_empty(),
-            "no machine_interface should exist before site-explorer reconciles for {mac}"
-        );
-    }
-    txn.commit().await?;
-
-    for (mac, ip, kind) in [
-        (machine_bmc_mac, machine_bmc_ip, "expected_machine BMC"),
-        (switch_bmc_mac, switch_bmc_ip, "expected_switch BMC"),
-        (
-            power_shelf_bmc_mac,
-            power_shelf_bmc_ip,
-            "expected_power_shelf BMC",
-        ),
-    ] {
-        carbide_site_explorer::try_preallocate_one(
-            &env.pool,
-            mac,
-            ip,
-            model::machine_interface::InterfaceType::Bmc,
-            kind,
-        )
-        .await;
-    }
-
-    let mut txn = env.pool.begin().await?;
-    for (mac, ip) in [
-        (machine_bmc_mac, machine_bmc_ip),
-        (switch_bmc_mac, switch_bmc_ip),
-        (power_shelf_bmc_mac, power_shelf_bmc_ip),
-    ] {
-        let after = db::machine_interface::find_by_mac_address(&mut *txn, mac).await?;
-        assert_eq!(after.len(), 1, "should be preallocated for {mac}");
-        assert!(
-            after[0].addresses.contains(&ip),
-            "preallocated row for {mac} should carry {ip}, got {:?}",
-            after[0].addresses,
-        );
-        assert_eq!(
-            after[0].interface_type,
-            model::machine_interface::InterfaceType::Bmc,
-            "BMC IPs should be preallocated with InterfaceType::Bmc, not Data ({mac})"
-        );
-    }
-
-    Ok(())
-}
-
-/// Site-explorer's reconciliation pass must materialize `ExpectedHostNic.fixed_ip`
-/// reservations too, not just BMC IPs.
-#[sqlx_test]
-async fn test_site_explorer_reconcile_preallocates_host_nic_fixed_ip(
-    pool: PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-
-    let bmc_mac: MacAddress = "AA:BB:CC:DD:E1:01".parse().unwrap();
-    let nic_mac: MacAddress = "AA:BB:CC:DD:E1:02".parse().unwrap();
-    let fixed_ip = "10.99.0.20";
-
-    let mut txn = env.pool.begin().await?;
-    db::expected_machine::create(
-        &mut txn,
-        ExpectedMachine {
-            id: None,
-            bmc_mac_address: bmc_mac,
-            data: ExpectedMachineData {
-                serial_number: "reconcile-hostnic-001".to_string(),
-                host_nics: vec![model::expected_machine::ExpectedHostNic {
-                    mac_address: nic_mac,
-                    nic_type: Some("onboard".into()),
-                    fixed_ip: Some(fixed_ip.into()),
-                    fixed_mask: None,
-                    fixed_gateway: None,
-                    primary: None,
-                }],
-                ..Default::default()
-            },
-        },
-    )
-    .await?;
-    txn.commit().await?;
-
-    let parsed_fixed_ip: IpAddr = fixed_ip.parse().unwrap();
-    carbide_site_explorer::try_preallocate_one(
-        &env.pool,
-        nic_mac,
-        parsed_fixed_ip,
-        model::machine_interface::InterfaceType::Data,
-        "expected_machine host NIC",
-    )
-    .await;
-
-    let mut txn = env.pool.begin().await?;
-    let nic_iface = db::machine_interface::find_by_mac_address(&mut *txn, nic_mac).await?;
-    assert_eq!(
-        nic_iface.len(),
-        1,
-        "host NIC interface should be preallocated"
-    );
-    assert!(
-        nic_iface[0].addresses.contains(&parsed_fixed_ip),
-        "preallocated host NIC interface should carry the fixed_ip"
-    );
-    assert_eq!(
-        nic_iface[0].interface_type,
-        model::machine_interface::InterfaceType::Data,
-        "host NIC preallocation should mark the interface as InterfaceType::Data, not Bmc"
-    );
-
-    Ok(())
-}
-
-/// Running `try_preallocate_one` twice for the same (mac, ip) must be a no-op the second time
-/// -- no new rows, no errors. This is the steady-state behavior since site-explorer iterates
-/// continuously and re-issues the same calls on every pass.
-#[sqlx_test]
-async fn test_site_explorer_reconcile_is_idempotent(
-    pool: PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-
-    let bmc_mac: MacAddress = "AA:BB:CC:DD:E2:01".parse().unwrap();
-    let bmc_ip: IpAddr = "10.99.0.30".parse().unwrap();
-
-    let mut txn = env.pool.begin().await?;
-    db::expected_machine::create(
-        &mut txn,
-        ExpectedMachine {
-            id: None,
-            bmc_mac_address: bmc_mac,
-            data: ExpectedMachineData {
-                serial_number: "reconcile-idem-001".to_string(),
-                bmc_ip_address: Some(bmc_ip),
-                ..Default::default()
-            },
-        },
-    )
-    .await?;
-    txn.commit().await?;
-
-    for _ in 0..2 {
-        carbide_site_explorer::try_preallocate_one(
-            &env.pool,
-            bmc_mac,
-            bmc_ip,
-            model::machine_interface::InterfaceType::Bmc,
-            "expected_machine BMC",
-        )
-        .await;
-    }
-
-    let mut txn = env.pool.begin().await?;
-    let interfaces = db::machine_interface::find_by_mac_address(&mut *txn, bmc_mac).await?;
-    assert_eq!(
-        interfaces.len(),
-        1,
-        "two preallocate calls should not create duplicate machine_interface rows"
-    );
-    assert!(interfaces[0].addresses.contains(&bmc_ip));
-
-    Ok(())
-}
-
-/// A per-entry conflict (e.g., two expected_machines configured with the same static IP -- a
-/// genuine operator misconfiguration) must not abort the whole reconciliation pass. The
-/// conflicting entry gets logged and skipped; the rest of the iteration continues.
-#[sqlx_test]
-async fn test_site_explorer_reconcile_tolerates_per_entry_conflicts(
-    pool: PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-
-    let mac_a: MacAddress = "AA:BB:CC:DD:E3:01".parse().unwrap();
-    let mac_b: MacAddress = "AA:BB:CC:DD:E3:02".parse().unwrap();
-    let mac_c: MacAddress = "AA:BB:CC:DD:E3:03".parse().unwrap();
-    let shared_ip: IpAddr = "10.99.0.40".parse().unwrap();
-    let ok_ip: IpAddr = "10.99.0.41".parse().unwrap();
-
-    let mut txn = env.pool.begin().await?;
-    // Two machines configured with the SAME bmc_ip_address -- the second will conflict at
-    // preallocate time. A third has its own valid IP and should still get preallocated.
-    for (mac, ip, sn) in [
-        (mac_a, shared_ip, "reconcile-conflict-a"),
-        (mac_b, shared_ip, "reconcile-conflict-b"),
-        (mac_c, ok_ip, "reconcile-conflict-c"),
-    ] {
-        db::expected_machine::create(
-            &mut txn,
-            ExpectedMachine {
-                id: None,
-                bmc_mac_address: mac,
-                data: ExpectedMachineData {
-                    serial_number: sn.to_string(),
-                    bmc_ip_address: Some(ip),
-                    ..Default::default()
-                },
-            },
-        )
-        .await?;
-    }
-    txn.commit().await?;
-
-    // try_preallocate_one swallows per-entry errors -- the conflict between mac_a and mac_b
-    // gets logged and the third call still succeeds.
-    for (mac, ip) in [(mac_a, shared_ip), (mac_b, shared_ip), (mac_c, ok_ip)] {
-        carbide_site_explorer::try_preallocate_one(
-            &env.pool,
-            mac,
-            ip,
-            model::machine_interface::InterfaceType::Bmc,
-            "expected_machine BMC",
-        )
-        .await;
-    }
-
-    let mut txn = env.pool.begin().await?;
-    // Exactly one of {mac_a, mac_b} won the race for the shared IP.
-    let a = db::machine_interface::find_by_mac_address(&mut *txn, mac_a).await?;
-    let b = db::machine_interface::find_by_mac_address(&mut *txn, mac_b).await?;
-    let winners = a.len() + b.len();
-    assert_eq!(
-        winners, 1,
-        "exactly one of the two conflicting MACs should have a preallocated row"
-    );
-    // The third, non-conflicting entry must still be preallocated.
-    let c = db::machine_interface::find_by_mac_address(&mut *txn, mac_c).await?;
-    assert_eq!(
-        c.len(),
-        1,
-        "non-conflicting expected_machine should still be preallocated despite the upstream conflict"
-    );
-    assert!(c[0].addresses.contains(&ok_ip));
-
-    Ok(())
-}
-
-/// Site-explorer's reconciliation pass must materialize the (nvos_mac, nvos_ip_address)
-/// pairing for expected switches, mirroring how it handles `bmc_ip_address` and the
-/// host-NIC `fixed_ip` paths. Calls `try_preallocate_one` directly the same way the
-/// expected_switches loop does, and verifies the resulting row carries the configured
-/// IP with `InterfaceType::Data`.
-#[sqlx_test]
-async fn test_site_explorer_reconcile_preallocates_nvos_ip(
-    pool: PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-
-    let bmc_mac: MacAddress = "AA:BB:CC:DD:E4:01".parse().unwrap();
-    let nvos_mac: MacAddress = "AA:BB:CC:DD:E4:02".parse().unwrap();
-    let nvos_ip: IpAddr = "10.99.0.50".parse().unwrap();
-
-    let mut txn = env.pool.begin().await?;
-    db::expected_switch::create(
-        &mut txn,
-        model::expected_switch::ExpectedSwitch {
-            expected_switch_id: None,
-            bmc_mac_address: bmc_mac,
-            nvos_mac_addresses: vec![nvos_mac],
-            bmc_username: "ADMIN".into(),
-            serial_number: "reconcile-nvos-001".into(),
-            bmc_password: "PASS".into(),
-            nvos_username: None,
-            nvos_password: None,
-            bmc_ip_address: None,
-            nvos_ip_address: Some(nvos_ip),
-            metadata: Metadata::default(),
-            rack_id: None,
-            bmc_retain_credentials: None,
-        },
-    )
-    .await?;
-    txn.commit().await?;
-
-    // Baseline: no NVOS interface yet.
-    let mut txn = env.pool.begin().await?;
-    let before = db::machine_interface::find_by_mac_address(&mut *txn, nvos_mac).await?;
-    assert!(
-        before.is_empty(),
-        "no machine_interface should exist before site-explorer reconciles for {nvos_mac}"
-    );
-    txn.commit().await?;
-
-    carbide_site_explorer::try_preallocate_one(
-        &env.pool,
-        nvos_mac,
-        nvos_ip,
-        model::machine_interface::InterfaceType::Data,
-        "expected_switch NVOS",
-    )
-    .await;
-
-    let mut txn = env.pool.begin().await?;
-    let after = db::machine_interface::find_by_mac_address(&mut *txn, nvos_mac).await?;
-    assert_eq!(
-        after.len(),
-        1,
-        "expected_switch NVOS should be preallocated for {nvos_mac}"
-    );
-    assert!(
-        after[0].addresses.contains(&nvos_ip),
-        "preallocated row should carry {nvos_ip}, got {:?}",
-        after[0].addresses,
-    );
-    assert_eq!(
-        after[0].interface_type,
-        model::machine_interface::InterfaceType::Data,
-        "NVOS IPs should be preallocated with InterfaceType::Data, not Bmc"
     );
 
     Ok(())
